Crate problem045

Problem 45 (Triangular, pentagonal and hexagonal)

Solution detail

Numbers which are triangular and pentagonal are found by solving the equation n × (n + 1) / 2 = m × (3m - 1) / 2. This can be put into the equivalent form:

(6m - 1)2 - 3 × (2n + 1)2 = -2

So we are searching for solutions (x, y) of the Pell-type equation x2 - 3y2 = -2 which satisfy:

x == 5 (mod 6)

y == 1 (mod 2)

Then m, n = (x + 1) / 6, (y - 1) / 2, and the figurate number is n × (n + 1) / 2

The solutions to this Pell equation are easy to find - the fundamental solution is (1, 1) and the generating solution is (2, 1), which means that we generate new solutions from old using the formula (x, y) ⟶ (2x + 3y, x + 2y).

For each solution we find, we can check whether the figurate number is also hexagonal, and if so, we have a solution.

Functions

answer
Solve the problem, returning the answer as a String

is_hexagonal
A function to check if a given number is hexagonal.

main

solve
Find the kth number which is simultaneously triangular, pentagonal and hexagonal.